Move end to end tests to separate test class, move end to end test files to separate folder, increase timer delay
