+++ src/test/resources/files/InventoryUUIDs.csv
@@ -1,0 +1,5 @@
+c8b50e3f-0446-429c-960e-03774b88223f,
+aae06d90-a8c2-4514-b227-5756f1f5f5d6,
+d5c7968c-17e7-4ab1-8aeb-3109e1b77c80,
+a5e9ccb3-737b-43b0-8f4a-f32a04c9ae16,
+c5d662af-b0be-4851-bb9c-de70bba3dfce
